delete: Удалены неиспользуемые методы delete: Удалены неиспользуемые классы feat: Дописана вся документация по всем методам refactor: Dao разложены по соответствующим папкам
