+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -56,6 +56,7 @@ public class Film {
      *
      * @see Film#Film(String, String, LocalDate, int)
      */
+
     @Autowired
     public Film(String name, String description, LocalDate releaseDate, int duration) {
         this.name = name;
