--- src/main/java/ru/yandex/practicum/filmorate/storage/dao/MpaDao.java
@@ -1,11 +1,0 @@
-package ru.yandex.practicum.filmorate.storage.dao;
-
-import ru.yandex.practicum.filmorate.model.Mpa;
-
-import java.util.List;
-
-public interface MpaDao {
-    Mpa getMpaById(Integer id);
-
-    List<Mpa> getListMpa();
-}
