+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/genre/GenreDaoImpl.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.dao;
+package ru.yandex.practicum.filmorate.storage.dao.genre;
 
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
