+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/mpa/MpaDao.java
@@ -1,0 +1,24 @@
+package ru.yandex.practicum.filmorate.storage.dao.mpa;
+
+import ru.yandex.practicum.filmorate.model.Mpa;
+
+import java.util.List;
+
+/**
+ * Интерфейс для работы с логикой касающиеся лайков, реализован в {@link MpaDaoImpl}
+ */
+public interface MpaDao {
+    /**
+     * Запрос рейтинга по идентификатору
+     *
+     * @param id айди пользователя, добавляющего лайк.
+     */
+    Mpa getMpaById(Integer id);
+
+    /**
+     * Запрос списка рейтингов.
+     *
+     * @return Возвращает список рейтингов.
+     */
+    List<Mpa> getListMpa();
+}
