+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/FriendMapper.java
@@ -6,7 +6,16 @@ import ru.yandex.practicum.filmorate.model.Friend;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+/**
+ * Маппер для реализации сущности Friend из данных полученных в БД
+ */
 public class FriendMapper implements RowMapper<Friend> {
+
+    /**
+     * Метод преобразования данных из БД в сущность
+     *
+     * @return возвращает сущность Friend
+     */
     @Override
     public Friend mapRow(ResultSet rs, int rowNum) throws SQLException {
         Friend friend = new Friend();
