+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/GenreMapper.java
@@ -6,7 +6,16 @@ import ru.yandex.practicum.filmorate.model.Genre;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+/**
+ * Маппер для реализации сущности Genre из данных полученных в БД
+ */
 public class GenreMapper implements RowMapper<Genre> {
+
+    /**
+     * Метод преобразования данных из БД в сущность
+     *
+     * @return возвращает сущность Genre
+     */
     @Override
     public Genre mapRow(ResultSet rs, int rowNum) throws SQLException {
         Genre genre = new Genre();
