--- src/main/java/ru/yandex/practicum/filmorate/storage/mapper/LikeMapper.java
@@ -1,17 +1,0 @@
-package ru.yandex.practicum.filmorate.storage.mapper;
-
-import org.springframework.jdbc.core.RowMapper;
-import ru.yandex.practicum.filmorate.model.Like;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public class LikeMapper implements RowMapper<Like> {
-    @Override
-    public Like mapRow(ResultSet rs, int rowNum) throws SQLException {
-        Like like = new Like();
-        like.setFilmId(rs.getLong("film_id"));
-        like.setUserId(rs.getLong("user_id"));
-        return like;
-    }
-}
