+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/MpaMapper.java
@@ -6,7 +6,16 @@ import ru.yandex.practicum.filmorate.model.Mpa;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+/**
+ * Маппер для реализации сущности Mpa из данных полученных в БД
+ */
 public class MpaMapper implements RowMapper<Mpa> {
+
+    /**
+     * Метод преобразования данных из БД в сущность
+     *
+     * @return возвращает сущность Mpa
+     */
     @Override
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
