+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/UserMapper.java
@@ -6,8 +6,16 @@ import ru.yandex.practicum.filmorate.model.User;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+/**
+ * Маппер для реализации сущности User из данных полученных в БД
+ */
 public class UserMapper implements RowMapper<User> {
 
+    /**
+     * Метод преобразования данных из БД в сущность
+     *
+     * @return возвращает сущность User
+     */
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         User user = new User();
